pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
replit = "^3.2.4"
Flask = "^2.1.3"
Flask-Assets = "^2.0"
Flask-Mobility = "^1.1.0"
waitress = "^2.1.2"
Pillow = "^9.3.0"
httpx = "^0.23.1"
prodict = "^0.8.18"
css-html-prettify = "^2.5.5"
beautifulsoup4 = "^4.11.1"
Flask-Caching = "^2.0.1"

[tool.poetry.dev-dependencies]
python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.0"}
debugpy = "^1.6.2"
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"